moving some code around to clean things up a bit
